+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-09 13:10:49
-// Commit: feat(perf): add bi-directional event bus handler for perf
+// Updated at: 2026-04-09 15:28:31
+// Commit: feat(perf): implement shimmer loading skeleton in perf
 
 class perfComponentState {
-  final String commitRef = '222';
-  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
-  final DateTime timestamp = DateTime.parse('2026-04-09T13:10:49+05:00');
+  final String commitRef = '227';
+  final String lastMessage = 'feat(perf): implement shimmer loading skeleton in perf';
+  final DateTime timestamp = DateTime.parse('2026-04-09T15:28:31+05:00');
 }